* 'master' of git://scm.dev.nokia.troll.no/qt/qt-water-team: (40 commits)
  Update Japanese translations.
  Updated Russian translation
  Update Korean translations to 4.8
  qmake: fix writing of incremental linker option for vcxproj files
  runonphone timestamps nanosecond formating set to 9 chars, 0 padded
  Put Qt translation files to rom on Symbian.
  Fix QDeclarativeEngine::setOfflineStoragePath() for Symbian
  Converting from double to qreal in gui
  Doc updates to installation, platform notes and symbian introduction.
  Fix XML schema validation failure.
  Fix construction races in QtNetwork
  Removing accidental use of double instead of qreal from QLine
  Fix incorrect QFileInfo permissions on windows
  Converting accidental use of doubles to qreal in declarative
  Copy and Cut are not available for some QML editors
  Ignore CcpuCan calls if input context is being destroyed
  Remove one local variable from QCoeFepInputContext
  Symbian: Fix native dialog with splitscreen VKB focus loss issue
  Create auto test for http HEAD request
  QNAM - fix poor performance of HEAD request with authentication
  ...

+++ src/declarative/util/qdeclarativespringanimation.cpp
@@ -161,17 +161,17 @@ bool QDeclarativeSpringAnimationPrivate::animate(const QDeclarativeProperty &pro
                 animation.velocity = animation.velocity + (spring * diff - damping * animation.velocity) / mass;
             else
                 animation.velocity = animation.velocity + spring * diff - damping * animation.velocity;
-            if (maxVelocity > 0.) {
+            if (maxVelocity > qreal(0.)) {
                 // limit velocity
                 if (animation.velocity > maxVelocity)
                     animation.velocity = maxVelocity;
                 else if (animation.velocity < -maxVelocity)
                     animation.velocity = -maxVelocity;
             }
-            animation.currentValue += animation.velocity * 16.0 / 1000.0;
+            animation.currentValue += animation.velocity * qreal(16.0) / qreal(1000.0);
             if (haveModulus) {
                 animation.currentValue = fmod(animation.currentValue, modulus);
-                if (animation.currentValue < 0.0)
+                if (animation.currentValue < qreal(0.0))
                     animation.currentValue += modulus;
             }
         }
@@ -195,7 +195,7 @@ bool QDeclarativeSpringAnimationPrivate::animate(const QDeclarativeProperty &pro
                 animation.currentValue = fmod(animation.currentValue, modulus);
         } else {
             animation.currentValue -= moveBy;
-            if (haveModulus && animation.currentValue < 0.0)
+            if (haveModulus && animation.currentValue < qreal(0.0))
                 animation.currentValue = fmod(animation.currentValue, modulus) + modulus;
         }
         if (lastTime - animation.start >= animation.duration) {
